ui/stats: merge to trx
