+++ data/common/ship/cfg/base_strings-de.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "\\{review}Verschiedenes",
         "SOUND_SETTINGS_TITLE": "Sound-Optionen",
         "SOUND_SETTINGS_VOLUME_TAB": "\\{review}Lautstärke",
+        "STATS_AMMO": "Munition Treffer/Verbraucht",
+        "STATS_AMMO_HITS": "Treffer",
+        "STATS_AMMO_USED": "Benötigte Munition",
         "STATS_ASSAULT_FINISH": "Zeit",
         "STATS_ASSAULT_NO_TIMES_SET": "Keine Zeiten gesetzt",
         "STATS_ASSAULT_TITLE": "Bestzeiten",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Bonusstatistiken",
+        "STATS_DEATHS": "Tode",
         "STATS_DETAIL_FMT": "%d von %d",
+        "STATS_DISTANCE_TRAVELLED": "Zurückgelegte Distanz",
         "STATS_FINAL_STATISTICS": "Finale Statistiken",
+        "STATS_KILLS": "Besiegte Gegner",
+        "STATS_LEVEL": "Level",
+        "STATS_MEDIPACKS_USED": "Benötigte Medi-Packs",
+        "STATS_NONE": "None",
+        "STATS_PICKUPS": "Pickups",
+        "STATS_SECRETS": "Gefundene Geheimnisse",
+        "STATS_TIME_TAKEN": "Benötigte Zeit",
     },
     "objects": {
         "flare": {
