+++ data/common/ship/cfg/base_strings-en-gb.json5
@@ -37,6 +37,7 @@
         "SETTING_WATER_COLOR_G_DESCRIPTION": "Green component of the water colour.",
         "SETTING_WATER_COLOR_R": "Water colour (R)",
         "SETTING_WATER_COLOR_R_DESCRIPTION": "Red component of the water colour.",
+        "STATS_DISTANCE_TRAVELLED": "Distance Travelled",
     },
     "objects": {}
 }
