+++ data/common/ship/cfg/base_strings-fr.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "\\{review}Divers",
         "SOUND_SETTINGS_TITLE": "Options sonores",
         "SOUND_SETTINGS_VOLUME_TAB": "\\{review}Volume",
+        "STATS_AMMO": "Tirs Réussis/Effectués",
+        "STATS_AMMO_HITS": "\\{review}Coups",
+        "STATS_AMMO_USED": "\\{review}Munitions utilisées",
         "STATS_ASSAULT_FINISH": "\\{review}Fin",
         "STATS_ASSAULT_NO_TIMES_SET": "\\{review}Aucun temps défini",
         "STATS_ASSAULT_TITLE": "\\{review}MEILLEURS TEMPS",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Statistiques bonus",
+        "STATS_DEATHS": "Morts",
         "STATS_DETAIL_FMT": "%d sur %d",
+        "STATS_DISTANCE_TRAVELLED": "Distance parcourue",
         "STATS_FINAL_STATISTICS": "Statistiques finales",
+        "STATS_KILLS": "Ennemis Tués",
+        "STATS_LEVEL": "\\{review}Niveau",
+        "STATS_MEDIPACKS_USED": "Trousses de soin utilisées",
+        "STATS_NONE": "\\{review}Aucun",
+        "STATS_PICKUPS": "Objets ramassés",
+        "STATS_SECRETS": "Secrets",
+        "STATS_TIME_TAKEN": "Temps écoulé",
     },
     "objects": {
         "flare": {
