+++ data/common/ship/cfg/base_strings-gd.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "\\{review}Iomlan",
         "SOUND_SETTINGS_TITLE": "Roghainnean Fuaime",
         "SOUND_SETTINGS_VOLUME_TAB": "\\{review}Tomhas",
+        "STATS_AMMO": "Buaidhean / Losgadh",
+        "STATS_AMMO_HITS": "Buaidhean",
+        "STATS_AMMO_USED": "Losgadh",
         "STATS_ASSAULT_FINISH": "Crìochnaich",
         "STATS_ASSAULT_NO_TIMES_SET": "Gun Àmanan Fhathast",
         "STATS_ASSAULT_TITLE": "NA H-ÀMANAN AS FHEÀRR",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Staitistig a Bharrachd",
+        "STATS_DEATHS": "Bàsan",
         "STATS_DETAIL_FMT": "%d de %d",
+        "STATS_DISTANCE_TRAVELLED": "Astar air a Shiubhal",
         "STATS_FINAL_STATISTICS": "Staitistig Dheireannach",
+        "STATS_KILLS": "Marbhain",
+        "STATS_LEVEL": "Ìre",
+        "STATS_MEDIPACKS_USED": "Pasganan Cleachdte",
+        "STATS_NONE": "Gin sam bith",
+        "STATS_PICKUPS": "Togailtean",
+        "STATS_SECRETS": "Dìomhaireachdan",
+        "STATS_TIME_TAKEN": "Ùine air a Ghabhail",
     },
     "objects": {
         "flare": {
