+++ data/common/ship/cfg/base_strings-it.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "Varie",
         "SOUND_SETTINGS_TITLE": "Opzioni audio",
         "SOUND_SETTINGS_VOLUME_TAB": "Volume",
+        "STATS_AMMO": "Munizioni a Segno/Usate",
+        "STATS_AMMO_HITS": "Colpi a Segno",
+        "STATS_AMMO_USED": "Munizioni Usate",
         "STATS_ASSAULT_FINISH": "Fine",
         "STATS_ASSAULT_NO_TIMES_SET": "Nessun tempo impostato",
         "STATS_ASSAULT_TITLE": "MIGLIORI TEMPI",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Statistiche Bonus",
+        "STATS_DEATHS": "Morti",
         "STATS_DETAIL_FMT": "%d di %d",
+        "STATS_DISTANCE_TRAVELLED": "Distanza Percorsa",
         "STATS_FINAL_STATISTICS": "Statistiche Finali",
+        "STATS_KILLS": "Uccisioni",
+        "STATS_LEVEL": "Livello",
+        "STATS_MEDIPACKS_USED": "Kit Medici Usati",
+        "STATS_NONE": "Nessuno",
+        "STATS_PICKUPS": "Oggetti",
+        "STATS_SECRETS": "Segreti Scoperti",
+        "STATS_TIME_TAKEN": "Tempo Impiegato",
     },
     "objects": {
         "flare": {
