+++ data/common/ship/cfg/base_strings-pl.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "Różne",
         "SOUND_SETTINGS_TITLE": "Ustawienia dźwięku",
         "SOUND_SETTINGS_VOLUME_TAB": "Głośność",
+        "STATS_AMMO": "Amunicja/trafienia",
+        "STATS_AMMO_HITS": "Trafienia",
+        "STATS_AMMO_USED": "Zużyta amunicja",
         "STATS_ASSAULT_FINISH": "Finisz",
         "STATS_ASSAULT_NO_TIMES_SET": "Brak czasów",
         "STATS_ASSAULT_TITLE": "NAJLEPSZE CZASY",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Statystyki bonusowe",
+        "STATS_DEATHS": "Śmierci",
         "STATS_DETAIL_FMT": "%d z %d",
+        "STATS_DISTANCE_TRAVELLED": "Przebyta odległość",
         "STATS_FINAL_STATISTICS": "Statystyki końcowe",
+        "STATS_KILLS": "Zabójstwa",
+        "STATS_LEVEL": "Poziom",
+        "STATS_MEDIPACKS_USED": "Zużyte apteczki",
+        "STATS_NONE": "Brak",
+        "STATS_PICKUPS": "Przedmioty",
+        "STATS_SECRETS": "Odnalezione sekrety",
+        "STATS_TIME_TAKEN": "Spędzony czas",
     },
     "objects": {
         "flare": {
