+++ data/common/ship/cfg/base_strings-ru.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "\\{review}Разное",
         "SOUND_SETTINGS_TITLE": "Настройки звука",
         "SOUND_SETTINGS_VOLUME_TAB": "\\{review}Громкость",
+        "STATS_AMMO": "Попадания/выстрелы",
+        "STATS_AMMO_HITS": "\\{review}Попадания",
+        "STATS_AMMO_USED": "\\{review}Использовано боеприпасов",
         "STATS_ASSAULT_FINISH": "\\{review}Завершить",
         "STATS_ASSAULT_NO_TIMES_SET": "\\{review}Время не установлено",
         "STATS_ASSAULT_TITLE": "\\{review}ЛУЧШИЕ ВРЕМЕНА",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Бонусная статистика",
+        "STATS_DEATHS": "Смертей",
         "STATS_DETAIL_FMT": "%d из %d",
+        "STATS_DISTANCE_TRAVELLED": "Пройденное расстояние",
         "STATS_FINAL_STATISTICS": "Финальная статистика",
+        "STATS_KILLS": "Убито",
+        "STATS_LEVEL": "Уровень",
+        "STATS_MEDIPACKS_USED": "Использовано аптечек",
+        "STATS_NONE": "\\{review}Нет",
+        "STATS_PICKUPS": "Предметы",
+        "STATS_SECRETS": "Секреты",
+        "STATS_TIME_TAKEN": "Время",
     },
     "objects": {
         "flare": {
