+++ data/common/ship/cfg/base_strings.json5
@@ -684,13 +684,25 @@
         "SOUND_SETTINGS_MISC_TAB": "Misc",
         "SOUND_SETTINGS_TITLE": "Sound Options",
         "SOUND_SETTINGS_VOLUME_TAB": "Volume",
+        "STATS_AMMO": "Ammo Hits/Used",
+        "STATS_AMMO_HITS": "Hits",
+        "STATS_AMMO_USED": "Ammo Used",
         "STATS_ASSAULT_FINISH": "Finish",
         "STATS_ASSAULT_NO_TIMES_SET": "No Times Set",
         "STATS_ASSAULT_TITLE": "BEST TIMES",
         "STATS_BASIC_FMT": "%d",
         "STATS_BONUS_STATISTICS": "Bonus Statistics",
+        "STATS_DEATHS": "Deaths",
         "STATS_DETAIL_FMT": "%d of %d",
+        "STATS_DISTANCE_TRAVELLED": "Distance Traveled",
         "STATS_FINAL_STATISTICS": "Final Statistics",
+        "STATS_KILLS": "Kills",
+        "STATS_LEVEL": "Level",
+        "STATS_MEDIPACKS_USED": "Health Packs Used",
+        "STATS_NONE": "None",
+        "STATS_PICKUPS": "Pickups",
+        "STATS_SECRETS": "Secrets Found",
+        "STATS_TIME_TAKEN": "Time Taken",
     },
     "objects": {
         "flare": {
