+++ data/tr1/ship/cfg/tr1/strings-de.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Fixt originale Fehler in Die Zisterne und Grab des Tihocan, bei denen Pflanzen-Sprites in Wassergebieten nicht animiert sind.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Fixt, dass die aufrecht stehend ausgeführten Prankenhiebe von Bären Lara nicht mehr verfehlen.",
         "STATS_AMMO": "Munition Treffer/Verbraucht",
+        "STATS_AMMO_HITS": "\\{review}TREFFER",
+        "STATS_AMMO_USED": "\\{review}VERBRAUCHTE MUNITION",
         "STATS_DEATHS": "Tode",
         "STATS_DISTANCE_TRAVELLED": "Zurückgelegte Distanz",
         "STATS_KILLS": "Besiegte Gegner",
         "STATS_LEVEL": "Level",
         "STATS_MEDIPACKS_USED": "Benötigte Medi-Packs",
+        "STATS_NONE": "\\{review}KEINE",
         "STATS_PICKUPS": "Pickups",
         "STATS_SECRETS": "Geheimnisse",
         "STATS_TIME_TAKEN": "Benötigte Zeit",
