+++ data/tr1/ship/cfg/tr1/strings-fr.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Corrige le fait que les sprites des végétations ne s'animent pas dans les zones aquatiques de La Citerne et de la Tombe de Tihocan.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Corrige l'attaque par coup de patte des ours pour qu'ils ne ratent pas Lara.",
         "STATS_AMMO": "TIRS RÉUSSIS/EFFECTUÉS",
+        "STATS_AMMO_HITS": "\\{review}COUPS",
+        "STATS_AMMO_USED": "\\{review}MUNITIONS UTILISÉES",
         "STATS_DEATHS": "MORTS",
         "STATS_DISTANCE_TRAVELLED": "DISTANCE PARCOURUE",
         "STATS_KILLS": "ENNEMIS TUÉS",
         "STATS_LEVEL": "\\{review}NIVEAU",
         "STATS_MEDIPACKS_USED": "TROUSSES DE SOIN UTILISÉES",
+        "STATS_NONE": "\\{review}AUCUN",
         "STATS_PICKUPS": "OBJETS RAMASSÉS",
         "STATS_SECRETS": "SECRETS",
         "STATS_TIME_TAKEN": "TEMPS ÉCOULÉ",
