+++ data/tr1/ship/cfg/tr1/strings-gd.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Càraichidh bugan san t-Siasarn is Uaigh Tihocan far nach gluais planntrais anns an uisge.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Càraichidh seo ionnsaigh a' mhathain gus nach caill e Lara.",
         "STATS_AMMO": "BUAIDHEAN / LOSGADH",
+        "STATS_AMMO_HITS": "\\{review}BUILEACHDAN",
+        "STATS_AMMO_USED": "\\{review}AMHACHD AIR A CHLEACHD",
         "STATS_DEATHS": "BÀSAN",
         "STATS_DISTANCE_TRAVELLED": "ASTAR AIR A SHIUBHAL",
         "STATS_KILLS": "MARBHAIN",
         "STATS_LEVEL": "ÌRE",
         "STATS_MEDIPACKS_USED": "PASGANAN CLEACHDTE",
+        "STATS_NONE": "\\{review}CHAN EIL",
         "STATS_PICKUPS": "TOGAILTEAN",
         "STATS_SECRETS": "DÌOMHAIREACHDAN",
         "STATS_TIME_TAKEN": "ÙINE AIR A GHABHAIL",
