+++ data/tr1/ship/cfg/tr1/strings-it.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Risolve i problemi preesistenti nella Cisterna e nella Tomba di Tihocan per cui gli sprite delle piante nelle aree acquatiche non si animano.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Risolve il problema per cui gli attacchi con la zampa degli orsi non colpiscono Lara.",
         "STATS_AMMO": "MUNIZIONI A SEGNO/USATE",
+        "STATS_AMMO_HITS": "\\{review}COLPI",
+        "STATS_AMMO_USED": "\\{review}MUNIZIONI USATE",
         "STATS_DEATHS": "MORTI",
         "STATS_DISTANCE_TRAVELLED": "DISTANZA PERCORSA",
         "STATS_KILLS": "UCCISIONI",
         "STATS_LEVEL": "LIVELLO",
         "STATS_MEDIPACKS_USED": "KIT MEDICI USATI",
+        "STATS_NONE": "\\{review}NESSUNO",
         "STATS_PICKUPS": "OGGETTI",
         "STATS_SECRETS": "SEGRETI",
         "STATS_TIME_TAKEN": "TEMPO IMPIEGATO",
