+++ data/tr1/ship/cfg/tr1/strings-pl.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Naprawia błędy oryginalnej gry w Rezerwuarze i Grobowcu Tihocana, gdzie glony w wodnych obszarach nie są animowane.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Naprawia atak niedźwiedzia, aby nie chybiał Lary.",
         "STATS_AMMO": "AMUNICJA/TRAFIENIA",
+        "STATS_AMMO_HITS": "\\{review}TRAFIENIA",
+        "STATS_AMMO_USED": "\\{review}ZUŻYTA AMUNICJA",
         "STATS_DEATHS": "ŚMIERCI",
         "STATS_DISTANCE_TRAVELLED": "PRZEBYTA ODLEGŁOŚĆ",
         "STATS_KILLS": "ZABÓJSTWA",
         "STATS_LEVEL": "POZIOM",
         "STATS_MEDIPACKS_USED": "UŻYTE APTECZKI",
+        "STATS_NONE": "\\{review}BRAK",
         "STATS_PICKUPS": "PRZEDMIOTY",
         "STATS_SECRETS": "SEKRETY",
         "STATS_TIME_TAKEN": "SPĘDZONY CZAS",
