+++ data/tr1/ship/cfg/tr1/strings-ru.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Исправляет изначальные проблемы в Цистерне и Гробнице Тихокана, из-за которых спрайты растений в водных зонах не анимировались.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Исправляет медвежью атаку, чтобы она не промахивалась по Ларе.",
         "STATS_AMMO": "Попадания/выстрелы",
+        "STATS_AMMO_HITS": "\\{review}ПОПАДАНИЯ",
+        "STATS_AMMO_USED": "\\{review}ИСПОЛЬЗОВАННЫЕ ПАТРОНЫ",
         "STATS_DEATHS": "Смертей",
         "STATS_DISTANCE_TRAVELLED": "Пройденное расстояние",
         "STATS_KILLS": "Убито",
         "STATS_LEVEL": "Уровень",
         "STATS_MEDIPACKS_USED": "Использовано аптечек",
+        "STATS_NONE": "\\{review}НЕТ",
         "STATS_PICKUPS": "Предметы",
         "STATS_SECRETS": "Секреты",
         "STATS_TIME_TAKEN": "Время",
