+++ data/tr1/ship/cfg/tr1/strings.json5
@@ -225,11 +225,14 @@
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Fixes original issues in The Cistern and Tomb of Tihocan where plant sprites in water areas do not animate.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Fixes the bear pat attack so it does not miss Lara.",
         "STATS_AMMO": "AMMO HITS/USED",
+        "STATS_AMMO_HITS": "HITS",
+        "STATS_AMMO_USED": "AMMO USED",
         "STATS_DEATHS": "DEATHS",
         "STATS_DISTANCE_TRAVELLED": "DISTANCE TRAVELED",
         "STATS_KILLS": "KILLS",
         "STATS_LEVEL": "LEVEL",
         "STATS_MEDIPACKS_USED": "HEALTH PACKS USED",
+        "STATS_NONE": "NONE",
         "STATS_PICKUPS": "PICKUPS",
         "STATS_SECRETS": "SECRETS",
         "STATS_TIME_TAKEN": "TIME TAKEN",
