+++ data/tr2/ship/cfg/tr2/strings-de.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "Harpune ausrüsen",
         "ENUM_INPUT_ROLE_EQUIP_M16": "M16-ausrüsten",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "Automatik-Pistolen ausrüsten",
-        "MISC_NONE": "None",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "Aktiviert die verwendung von dynamischer Beleuchtung für Schüsse und Explosionen.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "\\{review}Fixt originale Fehler in 40 Faden, Das Wrack der Maria Doria und An Deck, bei denen Pflanzen-Sprites in Wassergebieten nicht animiert sind.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Fixt, dass die aufrecht stehend ausgeführten Prankenhiebe von Bären Lara, im Addon 'Die Goldene Maske', nicht mehr verfehlen.",
-        "STATS_AMMO": "Munition Treffer/Verbraucht",
-        "STATS_AMMO_HITS": "Treffer",
-        "STATS_AMMO_USED": "Benötigte Munition",
-        "STATS_DEATHS": "Tode",
-        "STATS_DISTANCE_TRAVELLED": "Zurückgelegte Distanz",
-        "STATS_KILLS": "Besiegte Gegner",
-        "STATS_LEVEL": "Level",
-        "STATS_MEDIPACKS_USED": "Benötigte Medi-Packs",
-        "STATS_PICKUPS": "Pickups",
-        "STATS_SECRETS": "Gefundene Geheimnisse",
-        "STATS_TIME_TAKEN": "Benötigte Zeit",
     }
 }
