+++ data/tr2/ship/cfg/tr2/strings-fr.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "\\{review}Équiper le harpon",
         "ENUM_INPUT_ROLE_EQUIP_M16": "\\{review}Équiper le M16",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "\\{review}Équiper les pistolets automatiques",
-        "MISC_NONE": "\\{review}Aucun",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "\\{review}Permet de générer un éclairage dynamique pour les coups de feu et les explosions.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "\\{review}Corrige le fait que les sprites des végétations ne s'animent pas dans les zones aquatiques du 40 brasses, Épave du Maria Doria et Le Pont.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "\\{review}Corrige l'attaque de patte d'ours dans The Golden Mask pour qu'elle ne manque pas Lara.",
-        "STATS_AMMO": "Tirs Réussis/Effectués",
-        "STATS_AMMO_HITS": "\\{review}Coups",
-        "STATS_AMMO_USED": "\\{review}Munitions utilisées",
-        "STATS_DEATHS": "Morts",
-        "STATS_DISTANCE_TRAVELLED": "Distance parcourue",
-        "STATS_KILLS": "Ennemis Tués",
-        "STATS_LEVEL": "\\{review}Niveau",
-        "STATS_MEDIPACKS_USED": "Trousses de soin utilisées",
-        "STATS_PICKUPS": "Objets ramassés",
-        "STATS_SECRETS": "Secrets",
-        "STATS_TIME_TAKEN": "Temps écoulé",
     }
 }
