+++ data/tr2/ship/cfg/tr2/strings-gd.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "Uidheamaich Harpùn",
         "ENUM_INPUT_ROLE_EQUIP_M16": "Uidheamaich M16",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "Uidheamaich Piostalan Fèin-obrachail",
-        "MISC_NONE": "Gin sam bith",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "Cuiridh seo solais beòthail ri losgadh agus spreadhadh.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Càraichidh bugan ann an 40 Famhair, Long-bhriste na Maria Doria agus An Deic far nach gluais planntrais anns an uisge.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Càraichidh ionnsaigh a' mhathain anns a' Mhasg Òir gus nach caill e Lara.",
-        "STATS_AMMO": "Buaidhean / Losgadh",
-        "STATS_AMMO_HITS": "Buaidhean",
-        "STATS_AMMO_USED": "Losgadh",
-        "STATS_DEATHS": "Bàsan",
-        "STATS_DISTANCE_TRAVELLED": "Astar air a Shiubhal",
-        "STATS_KILLS": "Marbhain",
-        "STATS_LEVEL": "Ìre",
-        "STATS_MEDIPACKS_USED": "Pasganan Cleachdte",
-        "STATS_PICKUPS": "Togailtean",
-        "STATS_SECRETS": "Dìomhaireachdan",
-        "STATS_TIME_TAKEN": "Ùine air a Ghabhail",
     }
 }
