+++ data/tr2/ship/cfg/tr2/strings-it.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "Equipaggia Fucile Subacqueo",
         "ENUM_INPUT_ROLE_EQUIP_M16": "Equipaggia M16",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "Equipaggia Pistole Automatiche",
-        "MISC_NONE": "Nessuno",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "Abilita l'illuminazione dinamica per spari ed esplosioni.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Risolve i problemi preesistenti nei livelli 'In Profondità', 'Relitto della Maria Doria' e 'Sul Ponte' per cui le animazioni degli sprite delle piante nelle aree acquatiche non funzionano.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Risolve il problema nel pacchetto di espansione La Maschera Dorata per cui gli attacchi con la zampa degli orsi non colpiscono Lara.",
-        "STATS_AMMO": "Munizioni a Segno/Usate",
-        "STATS_AMMO_HITS": "Colpi a Segno",
-        "STATS_AMMO_USED": "Munizioni Usate",
-        "STATS_DEATHS": "Morti",
-        "STATS_DISTANCE_TRAVELLED": "Distanza Percorsa",
-        "STATS_KILLS": "Uccisioni",
-        "STATS_LEVEL": "Livello",
-        "STATS_MEDIPACKS_USED": "Kit Medici Usati",
-        "STATS_PICKUPS": "Oggetti",
-        "STATS_SECRETS": "Segreti Scoperti",
-        "STATS_TIME_TAKEN": "Tempo Impiegato",
     }
 }
