+++ data/tr2/ship/cfg/tr2/strings-pl.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "Wyposaż wyrzutnię harpunów",
         "ENUM_INPUT_ROLE_EQUIP_M16": "Wyposaż M16",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "Wyposaż pistolety automatyczne",
-        "MISC_NONE": "Brak",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "Włącza dynamiczne oświetlenie przy strzelaniu z broni i eksplozjach.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Przywraca brakującą animację glonów w podwodnych poziomach.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Naprawia atak niedźwiedzia w dodatku Złota Maska, aby nie chybiał Lary.",
-        "STATS_AMMO": "Amunicja/trafienia",
-        "STATS_AMMO_HITS": "Trafienia",
-        "STATS_AMMO_USED": "Zużyta amunicja",
-        "STATS_DEATHS": "Śmierci",
-        "STATS_DISTANCE_TRAVELLED": "Przebyta odległość",
-        "STATS_KILLS": "Zabójstwa",
-        "STATS_LEVEL": "Poziom",
-        "STATS_MEDIPACKS_USED": "Zużyte apteczki",
-        "STATS_PICKUPS": "Przedmioty",
-        "STATS_SECRETS": "Odnalezione sekrety",
-        "STATS_TIME_TAKEN": "Spędzony czas",
     }
 }
