+++ data/tr2/ship/cfg/tr2/strings.json5
@@ -340,20 +340,8 @@
         "ENUM_INPUT_ROLE_EQUIP_HARPOON": "Equip Harpoon Gun",
         "ENUM_INPUT_ROLE_EQUIP_M16": "Equip M16",
         "ENUM_INPUT_ROLE_EQUIP_MAGNUMS": "Equip Automatic Pistols",
-        "MISC_NONE": "None",
         "SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION": "Enables dynamic lighting to be generated for gunshots and explosions.",
         "SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION": "Fixes original issues in 40 Fathoms, Wreck of the Maria Doria and The Deck where plant sprites in water areas do not animate.",
         "SETTING_FIX_BEAR_AI_DESCRIPTION": "Fixes the bear pat attack in The Golden Mask so it does not miss Lara.",
-        "STATS_AMMO": "Ammo Hits/Used",
-        "STATS_AMMO_HITS": "Hits",
-        "STATS_AMMO_USED": "Ammo Used",
-        "STATS_DEATHS": "Deaths",
-        "STATS_DISTANCE_TRAVELLED": "Distance Traveled",
-        "STATS_KILLS": "Kills",
-        "STATS_LEVEL": "Level",
-        "STATS_MEDIPACKS_USED": "Health Packs Used",
-        "STATS_PICKUPS": "Pickups",
-        "STATS_SECRETS": "Secrets Found",
-        "STATS_TIME_TAKEN": "Time Taken",
     }
 }
