+++ src/libtrx/include/libtrx/game/game_string.def
@@ -706,3 +706,16 @@ GS_DEFINE(GRAPHIC_SETTINGS_TITLE,                   "Graphic Options")
 GS_DEFINE(GRAPHIC_SETTINGS_VISUALS_TAB,             "Visuals")
 GS_DEFINE(GRAPHIC_SETTINGS_UI_TAB,                  "UI")
 GS_DEFINE(GRAPHIC_SETTINGS_RENDERING_TAB,           "Rendering")
+
+GS_DEFINE(STATS_LEVEL,              "Level")
+GS_DEFINE(STATS_TIME_TAKEN,         "Time Taken")
+GS_DEFINE(STATS_AMMO,               "Ammo Hits/Used")
+GS_DEFINE(STATS_AMMO_USED,          "Ammo Used")
+GS_DEFINE(STATS_AMMO_HITS,          "Hits")
+GS_DEFINE(STATS_KILLS,              "Kills")
+GS_DEFINE(STATS_PICKUPS,            "Pickups")
+GS_DEFINE(STATS_DEATHS,             "Deaths")
+GS_DEFINE(STATS_MEDIPACKS_USED,     "Health Packs Used")
+GS_DEFINE(STATS_DISTANCE_TRAVELLED, "Distance Traveled")
+GS_DEFINE(STATS_SECRETS,            "Secrets Found")
+GS_DEFINE(STATS_NONE,               "None")
