+++ src/tr1/game/game_string.def
@@ -15,4 +15,7 @@ GS_DEFINE(STATS_TIME_TAKEN, "TIME TAKEN")
 GS_DEFINE(STATS_AMMO, "AMMO HITS/USED")
 GS_DEFINE(STATS_DISTANCE_TRAVELLED, "DISTANCE TRAVELED")
 GS_DEFINE(STATS_MEDIPACKS_USED, "HEALTH PACKS USED")
+GS_DEFINE(STATS_AMMO_USED, "AMMO USED")
+GS_DEFINE(STATS_AMMO_HITS, "HITS")
+GS_DEFINE(STATS_NONE, "NONE")
 GS_DEFINE(MISC_EMPTY_SLOT_FMT, "- EMPTY SLOT %d -")
