+++ src/tr2/game/game_string.def
@@ -9,16 +9,3 @@ GS_DEFINE(ENUM_INPUT_ROLE_EQUIP_MAGNUMS, "Equip Automatic Pistols")
 GS_DEFINE(SETTING_FIX_ANIMATED_SPRITES_DESCRIPTION,  "Fixes original issues in 40 Fathoms, Wreck of the Maria Doria and The Deck where plant sprites in water areas do not animate.")
 GS_DEFINE(SETTING_FIX_BEAR_AI_DESCRIPTION,           "Fixes the bear pat attack in The Golden Mask so it does not miss Lara.")
 GS_DEFINE(SETTING_ENABLE_GUN_LIGHTING_DESCRIPTION,   "Enables dynamic lighting to be generated for gunshots and explosions.")
-
-GS_DEFINE(STATS_LEVEL, "Level")
-GS_DEFINE(STATS_TIME_TAKEN, "Time Taken")
-GS_DEFINE(STATS_SECRETS, "Secrets Found")
-GS_DEFINE(STATS_AMMO, "Ammo Hits/Used")
-GS_DEFINE(STATS_AMMO_USED, "Ammo Used")
-GS_DEFINE(STATS_AMMO_HITS, "Hits")
-GS_DEFINE(STATS_KILLS, "Kills")
-GS_DEFINE(STATS_PICKUPS, "Pickups")
-GS_DEFINE(STATS_DEATHS, "Deaths")
-GS_DEFINE(STATS_MEDIPACKS_USED, "Health Packs Used")
-GS_DEFINE(STATS_DISTANCE_TRAVELLED, "Distance Traveled")
-GS_DEFINE(MISC_NONE, "None")
