--- src/tr2/game/ui/dialogs/stats.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/ui/dialogs/stats.h>
