+++ src/tr2/meson.build
@@ -94,7 +94,6 @@ sources = [
   'game/objects/common.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
-  'game/ui/dialogs/stats.c',
   resources,
 ]
